OverView: 增加商品库存和销售关联 Type: feature BUG: NA Description: please enter the long description

+++ app/src/main/java/com/sgm/a3dshop/ui/products/ProductDetailViewModel.kt
@@ -29,7 +29,7 @@ class ProductDetailViewModel(
 
     private fun loadProduct() {
         viewModelScope.launch {
-            val loadedProduct = productDao.getProductById(productId)
+            val loadedProduct = productDao.getProductById(productId.toInt())
             Log.d(TAG, "从数据库加载产品数据:")
             Log.d(TAG, "- ID: ${loadedProduct.id}")
             Log.d(TAG, "- 名称: ${loadedProduct.name}")
